Fix float32 to float64 on Beer model

+++ model/beer.go
@@ -7,13 +7,13 @@ type Beer struct {
 	FirstBrewed      string       `json:"first_brewed"`
 	Description      string       `json:"description"`
 	ImageUrl         string       `json:"image_url"`
-	Abv              float32      `json:"abv"`
-	Ibu              float32      `json:"ibu"`
-	TargetFg         float32      `json:"target_fg"`
-	TargetOg         float32      `json:"target_og"`
-	Ebc              float32      `json:"ebc"`
-	Srm              float32      `json:"srm"`
-	Ph               float32      `json:"ph"`
+	Abv              float64      `json:"abv"`
+	Ibu              float64      `json:"ibu"`
+	TargetFg         float64      `json:"target_fg"`
+	TargetOg         float64      `json:"target_og"`
+	Ebc              float64      `json:"ebc"`
+	Srm              float64      `json:"srm"`
+	Ph               float64      `json:"ph"`
 	AttenuationValue int          `json:"attenuation_value"`
 	Volume           *BeerVolume  `json:"volume"`
 	FoodPairing      *FoodPairing `json:"food_pairing"`
